Output progress status only in text mode
