+++ cli/core/upgrade.go
@@ -24,7 +24,6 @@ import (
 	"github.com/arduino/arduino-cli/cli"
 	"github.com/arduino/arduino-cli/commands/core"
 	"github.com/arduino/arduino-cli/common/formatter"
-	"github.com/arduino/arduino-cli/output"
 	"github.com/arduino/arduino-cli/rpc"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -56,7 +55,7 @@ func runUpgradeCommand(cmd *cobra.Command, args []string) {
 			PlatformPackage: platformRef.Package,
 			Architecture:    platformRef.PlatformArchitecture,
 			Version:         platformRef.PlatformVersion.String(),
-		}, output.DownloadProgressBar(), output.NewTaskProgressCB())
+		}, cli.OutputProgressBar(), cli.OutputTaskProgress())
 		if err != nil {
 			formatter.PrintError(err, "Error during upgrade")
 			os.Exit(cli.ErrGeneric)
